# ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package ports

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// Note defines model for Note.
type Note struct {
	Description string             `json:"description"`
	Time        time.Time          `json:"time"`
	Title       string             `json:"title"`
	Uuid        openapi_types.UUID `json:"uuid"`
}

// Notes defines model for Notes.
type Notes struct {
	Notes []Note `json:"notes"`
}

// PostNote defines model for PostNote.
type PostNote struct {
	Description string `json:"description"`
	Title       string `json:"title"`
}

// CreateNoteJSONRequestBody defines body for CreateNote for application/json ContentType.
type CreateNoteJSONRequestBody = PostNote

// UpdateNoteJSONRequestBody defines body for UpdateNote for application/json ContentType.
type UpdateNoteJSONRequestBody = PostNote